# perf(core): handle Buffer allocation in `hashUtils.checksumFile`